drf/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
boto3==1.35.90
botocore==1.35.90
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Deprecated==1.2.15
Django==5.1.4
django_celery_beat==2.7.0
django-cors-headers==4.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
minio==7.2.13
multidict==6.1.0
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.0
yarl==1.18.3
